+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - Install lsof in the `cli` profile.
 - Install file in the `cli` profile.
 - Install the OpenSSH client in the `cli` profile.
+- Install DNS lookup utilities (`dig`, `host`, `nslookup`) in the `cli` profile.
+- Install iproute2 (`ip`) in the `cli` profile.
+- Install ping in the `cli` profile.
 
 ## [1.6.0] - 2026-08-10